Feature: Add support for handling connection retries
Dag Wieers originally submitted this work as a PR in Oct 2018. Due to
the amount of time that has passed, the PR needs to be rebased. Dag
has not been able to get back to this.
I am therefore submitting a new PR with his proposed changes, and added
some tests as well.

Original PR from Dag:
https://github.com/diyan/pywinrm/pull/174